Fixes for the unused bot.
