+++ spyder/widgets/tests/test_editor.py
@@ -8,36 +8,11 @@
 Tests for editor.py
 """
 
-# Standard library imports
-try:
-    from unittest.mock import Mock
-except ImportError:
-    from mock import Mock # Python 2
-
 # Third party imports
 import pytest
 
 # Local imports
-from spyder.widgets.editor import EditorStack
-from spyder.widgets.findreplace import FindReplace
-
-def setup_editor(qtbot):
-    """
-    Set up EditorStack with CodeEditor containing some Python code.
-    The cursor is at the empty line below the code.
-    Returns tuple with EditorStack and CodeEditor.
-    """
-    text = ('a = 1\n'
-            'print(a)\n'
-            '\n'
-            'x = 2')  # a newline is added at end
-    editorStack = EditorStack(None, [])
-    editorStack.set_introspector(Mock())
-    editorStack.set_find_widget(FindReplace(editorStack))
-    editorStack.set_io_actions(Mock(), Mock(), Mock(), Mock())
-    finfo = editorStack.new('foo.py', 'utf-8', text)
-    qtbot.addWidget(editorStack)
-    return editorStack, finfo.editor
+from spyder.utils.fixtures import setup_editor
 
 def test_run_top_line(qtbot):
     editorStack, editor = setup_editor(qtbot)
